\name{PlotSolarRadiation}
\alias{PlotSolarRadiation}
\title{Plot the Fate of Incoming Global Solar Radiation}
\description{
  Create two plots indicating what happens to the incoming solar
  radiation based on the FLake output \code{data.table}.
}
\usage{
PlotSolarRadiation(flakeresult)
}
\arguments{
  \item{flakeresult}{the \code{data.frame} object from
    \code{flaker::runflake}}
}
%% \seealso{
%%   \code{\link{runflake}}.
%% }

